Store multiple player states in the game scene
